chatnest/backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv 
firebase-admin 
pyjwt
httpx
requests 